feat: merge test refactoring and production fixes from tests-to-100
- Add respx-based HTTP edge mocking for tests
- Reorganize test fixtures into subdirectories

Production bug fixes:
- fix(fileio): correct dedupe logic (return False instead of True)
- fix(schema): change Attachment.id from INTEGER to BIGINT for snowflake IDs
- fix(database): add PostgreSQL timezone configuration for UTC handling
- fix(types): add return type hints to __getattr__ methods
- fix(logging): convert debug prints to logger.debug() calls

Partial migration progress: 2/305 test functions migrated from mock-based to edge-mocking patterns.
Test migration continues with comprehensive TODO tracking in STASH_TEST_MIGRATION_TODO.md.

+++ metadata/attachment.py
@@ -87,7 +87,7 @@ class Attachment(Base):
 
     __tablename__ = "attachments"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
     postId: Mapped[int | None] = mapped_column(
         BigInteger, ForeignKey("posts.id"), nullable=True
     )
